feat(hosts.thinkpad-e580): disable nix-serve
Nix is not reliable if a binary cache is down. My cluster should still be able to function regardless of whether my laptop is on or not, so I'm going to disable my laptop's binary cache.

I'm doing this now because I'm finally getting some stuff running that does not need any action from me (and does not need my laptop on). See jyooru/nix-minecraft-servers#49.